move layer instance
